+++ source/adios2/CMakeLists.txt
@@ -234,8 +234,6 @@ if(ADIOS2_HAVE_MPI)
   target_link_libraries(adios2 PUBLIC MPI::MPI_C)
 endif()
 
-target_sources(adios2 PRIVATE helper/mpidummy.cpp helper/mpiwrap.cpp)
-
 if(ADIOS2_HAVE_HDF5)
   if(HDF5_C_INCLUDE_DIRS)
     target_include_directories(adios2 PRIVATE ${HDF5_C_INCLUDE_DIRS})
