+++ source/adios2/engine/sst/SstReader.cpp
@@ -22,7 +22,7 @@
 #ifdef ADIOS2_HAVE_MPI
 #include "adios2/helper/adiosCommMPI.h"
 #else
-#include "adios2/helper/mpiwrap.h"
+#include "adios2/toolkit/sst/mpiwrap.h"
 #endif
 
 namespace adios2
