+++ source/adios2/toolkit/sst/dp_interface.h
@@ -4,7 +4,7 @@
 #include <sst_data.h>
 
 #include "adios2/common/ADIOSConfig.h"
-#include "adios2/helper/mpiwrap.h"
+#include "mpiwrap.h"
 #include <evpath.h>
 
 /*!
